Use custom bridge for test accounts
