+++ modules/web3/helpers/getContracts.ts
@@ -2,8 +2,11 @@ import { ethers, Signer } from 'ethers';
 import { getGoerliSdk, getMainnetSdk, GoerliSdk, MainnetSdk } from '@dethcrypto/eth-sdk-client';
 
 import { Web3Provider } from '@ethersproject/providers';
-import { CHAIN_INFO, SupportedNetworks } from '../constants/networks';
+import { CHAIN_INFO, DEFAULT_NETWORK, SupportedNetworks } from '../constants/networks';
 import { ZERO_ADDRESS } from 'modules/web3/constants/addresses';
+import { SupportedChainId } from '../constants/chainID';
+import { getRPCFromChainID } from './getRPC';
+import { chainIdToNetworkName } from './chain';
 
 export type EthSdk = MainnetSdk | GoerliSdk;
 
@@ -19,13 +22,15 @@ const sdks: Sdks = {
 
 // this name doesn't feel right, maybe getSdk? or getContractLibrary?
 export const getContracts = (
-  chainId?: number,
+  chainId?: SupportedChainId,
   library?: Web3Provider,
   account?: string | undefined | null
 ): EthSdk => {
-  const network = chainId ? CHAIN_INFO[chainId].network : SupportedNetworks.MAINNET;
-  const provider = ethers.getDefaultProvider(network);
+  const provider = ethers.getDefaultProvider(getRPCFromChainID(chainId || 1));
+  const network = chainId ? chainIdToNetworkName(chainId) : DEFAULT_NETWORK;
 
+  // Map goerlifork to goerli contracts
+  const sdkNetwokrKey = network === SupportedNetworks.GOERLIFORK ? SupportedNetworks.GOERLI : network;
   /* 
   A read-only signer, when an API requires a Signer as a parameter, but it is known only read-only operations will be carried.
   https://docs.ethers.io/v5/api/signer/#VoidSigner 
@@ -36,5 +41,5 @@ export const getContracts = (
   const signer =
     account && library ? library.getSigner(account) : new ethers.VoidSigner(ZERO_ADDRESS, provider);
 
-  return sdks[network](signer);
+  return sdks[sdkNetwokrKey](signer);
 };
