+++ modules/web3/hooks/useWindowBindings.tsx
@@ -1,40 +1,36 @@
 import { useWeb3React } from '@web3-react/core';
-import { useEffect, useState } from 'react';
-import { PrivateKeyConnector } from '../connectors/PrivateKeyConnector';
-
-type PrivateKeyAccount = {
-  address: string;
-  key: string;
-};
+import { Wallet } from 'ethers';
+import { useEffect } from 'react';
+import { SupportedChainId } from '../constants/chainID';
+import { JsonRpcProvider } from '@ethersproject/providers';
+import { InjectedConnector } from '@web3-react/injected-connector';
+import { CustomizedBridge } from '../connectors/CustomizedBridge';
 
 export function useWindowBindings(): void {
-  const [account, setAccount] = useState<PrivateKeyAccount>({
-    address: '',
-    key: ''
-  });
-
   const context = useWeb3React();
   // Define a window function that changes the account for testing purposes
   useEffect(() => {
     if (typeof window !== 'undefined') {
       window.setAccount = (address: string, key: string) => {
-        setAccount({
-          address,
-          key
-        });
+        if (address && key) {
+          const rpcUrl = 'http://localhost:8545';
+          const provider = new JsonRpcProvider(rpcUrl, SupportedChainId.GOERLIFORK);
+          const signer = new Wallet(key, provider);
+
+          const bridge = new CustomizedBridge(signer, provider);
+          bridge.setAddress(address);
+          window.ethereum = bridge;
+
+          context.activate(
+            new InjectedConnector({
+              supportedChainIds: [SupportedChainId.GOERLIFORK]
+            }),
+            err => {
+              console.log('error', err);
+            }
+          );
+        }
       };
     }
   }, []);
-
-  useEffect(() => {
-    const changeMakerAccount = async () => {
-      if (account.address && account.key) {
-        const connector = new PrivateKeyConnector(account.key, account.address);
-
-        context.activate(connector);
-      }
-    };
-
-    changeMakerAccount();
-  }, [account]);
 }
